Add files via upload
These Python scripts are for the bulk of the analysis of what goes into the paper. This alone is enough to find Toomre Q in a given galaxy. This was specifically for the galaxy NGC 0524. The only things that change for the different galaxies are physical parameters and file locations.